Improve message from `refute` on failure
The previous implementation was to `assert(!value)` which led to
confusing failure messages when using `refute` directly.

For custom failure messages, `!value` is always `false` on failure, so
`false` is passed to the custom message block where the caller might
instead expect the value that caused the failure.

Similarly, the default failure message is always "expected false to be
truthy" when really the expectation is that something truthy is falsy.